+++ pyvo/dal/sia.py
@@ -706,7 +706,7 @@ class SIARecord(SodaRecordMixin, DatalinkRecordMixin, Record):
     @property
     def dateobs(self):
         """
-        the modified Julien date (MJD) of the mid-point of the
+        the modified Julian date (MJD) of the mid-point of the
         observational data that went into the image,
         as an astropy.time.Time instance
         """
@@ -807,13 +807,24 @@ class SIARecord(SodaRecordMixin, DatalinkRecordMixin, Record):
         """
         the astropy unit used to represent spectral values.
         """
-        sia = self.getbyucd("VOX:BandPass_Unit", decode=True)
+        sia_unit = self.getbyucd("VOX:BandPass_Unit", decode=True)
 
-        if sia:
-            return Unit(sia)
-        else:
-            # dimensionless
-            return Unit("")
+        # the standard says this should be `"meters", "hertz", and "keV"',
+        # which in practice everyone has ignored.  Still, let's
+        # translate the standard terms.  Somewhat more dangerously,
+        # we replace a missing unit with metres; in theory, we should
+        # reject anything but the three terms above, but then 99%
+        # of SIA services would break.  And without the assumption of
+        # a metre default, 20% of 2024 SIA1 services would have unusable
+        # bandpasses.
+
+        astropy_unit = {
+            None: "m",
+            "": "m",
+            "meters": "m",
+            "hertz": "Hz"}.get(sia_unit, sia_unit)
+
+        return Unit(astropy_unit)
 
     @property
     def bandpass_refvalue(self):
@@ -821,24 +832,27 @@ class SIARecord(SodaRecordMixin, DatalinkRecordMixin, Record):
         the characteristic (reference) wavelength, frequency or energy
         for the bandpass model, as an astropy Quantity of bandpass_unit
         """
-        return Quantity(
-            self.getbyucd("VOX:BandPass_RefValue"), self.bandpass_unit)
+        value = self.getbyucd("VOX:BandPass_RefValue")
+        if value is not None and self.bandpass_unit:
+            return Quantity(value, self.bandpass_unit)
 
     @property
     def bandpass_hilimit(self):
         """
         the upper limit of the bandpass as astropy Quantity in bandpass_unit
         """
-        return Quantity(
-            self.getbyucd("VOX:BandPass_HiLimit"), self.bandpass_unit)
+        value = self.getbyucd("VOX:BandPass_HiLimit")
+        if value is not None and self.bandpass_unit:
+            return Quantity(value, self.bandpass_unit)
 
     @property
     def bandpass_lolimit(self):
         """
         the lower limit of the bandpass as astropy Quantity in bandpass_unit
         """
-        return Quantity(
-            self.getbyucd("VOX:BandPass_LoLimit"), self.bandpass_unit)
+        value = self.getbyucd("VOX:BandPass_LoLimit")
+        if value is not None and self.bandpass_unit:
+            return Quantity(value, self.bandpass_unit)
 
     # Processig Metadata
     @property