+++ pyvo/registry/rtcons.py
@@ -519,11 +519,17 @@ class Datamodel(SubqueriedConstraint):
     one of several well-known data models; the SQL produced depends
     on the data model identifier.
 
+    Records returned with these constraints will have a TAP (or
+    auxiliary TAP) capability.
+
     Known data models at this point include:
 
     * obscore -- generic observational data
     * epntap -- solar system data
     * regtap -- the VO registry.
+    * obscore-new -- the table-based new-style obscore discovery.  Don't
+      use in code built to last: this will become normal obscore
+      when we have migrated the VO.
 
     DM names are matched case-insensitively here mainly for
     historical reasons.
@@ -532,7 +538,7 @@ class Datamodel(SubqueriedConstraint):
 
     # if you add to this list, you have to define a method
     # _make_<dmname>_constraint.
-    _known_dms = {"obscore", "epntap", "regtap"}
+    _known_dms = {"obscore", "epntap", "regtap", "obscore_new"}
 
     def __init__(self, dmname):
         """
@@ -556,6 +562,14 @@ class Datamodel(SubqueriedConstraint):
         return "rr.res_detail", (
             "detail_xpath = '/capability/dataModel/@ivo-id'"
             f" AND 1 = ivo_nocasematch(detail_value, '{obscore_pat}')")
+
+    def _make_obscore_new_constraint(self):
+        return "rr.res_table NATURAL JOIN rr.resource", (
+            "table_utype LIKE 'ivo://ivoa.net/std/obscore#table-1.%'"
+            # Only use catalogresource-typed records to keep out
+            # full TAP services that may have the table in their
+            # tablesets.
+            " AND res_type = 'vs:catalogresource'")
 
     def _make_epntap_constraint(self):
         # we include legacy, pre-IVOA utypes for matches; lowercase
@@ -683,7 +697,8 @@ class Spatial(SubqueriedConstraint):
 
     takes_sequence = True
 
-    def __init__(self, geom_spec, order=6, intersect="covers"):
+    def __init__(self, geom_spec,
+            *, order=6, intersect="covers", inclusive=False):
         """
 
         Parameters
@@ -708,8 +723,13 @@ class Spatial(SubqueriedConstraint):
             that completely cover the *geom_spec* region, 'enclosed' for services
             completely enclosed in the region and 'overlaps' for services which
             coverage intersect the region.
-
+        inclusive : bool, optional
+            Normally, this constraint will remove all resources that do
+            not declare their spatial coverage.  Pass inclusive=True to
+            retain these.
         """
+        self.inclusive = inclusive
+
         def tomoc(s):
             return _AsIs("MOC({}, {})".format(order, s))
 
@@ -763,15 +783,21 @@ class Spatial(SubqueriedConstraint):
         # MOC-based geometries but does not have a MOC function.
         if not service.get_tap_capability().get_adql().get_feature(
                 "ivo://org.gavo.dc/std/exts#extra-adql-keywords", "MOC"):
-            raise RegTAPFeatureMissing("Current RegTAP service does not support MOC.")
+            raise RegTAPFeatureMissing(
+                "Current RegTAP service does not support MOC.")
 
         # We should compare case-insensitively here, but then we don't
         # with delimited identifiers -- in the end, that would have to
         # be handled in dal.vosi.VOSITables.
         if "rr.stc_spatial" not in service.tables:
-            raise RegTAPFeatureMissing("stc_spatial missing on current RegTAP service")
+            raise RegTAPFeatureMissing(
+                "stc_spatial missing on current RegTAP service")
 
-        return super().get_search_condition(service)
+        cond = super().get_search_condition(service)
+        if self.inclusive:
+            return cond[:-1]+" OR coverage IS NULL)"
+        else:
+            return cond
 
 
 class Spectral(SubqueriedConstraint):
@@ -811,7 +837,7 @@ class Spectral(SubqueriedConstraint):
 
     takes_sequence = True
 
-    def __init__(self, spec):
+    def __init__(self, spec, *, inclusive=False):
         """
 
         Parameters
@@ -822,7 +848,14 @@ class Spectral(SubqueriedConstraint):
             in which case the argument is interpreted as an interval.
             All resources *overlapping* the interval are returned.
             Plain floats are interpreted as messenger energy in Joule.
+
+        inclusive : bool, optional
+            Normally, this constraint will remove all resources that do
+            not declare their spectral coverage.  Pass inclusive=True to
+            retain these.
         """
+        self.inclusive = inclusive
+
         if isinstance(spec, tuple):
             self._fillers = {
                 "spec_lo": self._to_joule(spec[0]),
@@ -865,9 +898,15 @@ class Spectral(SubqueriedConstraint):
 
     def get_search_condition(self, service):
         if "rr.stc_spectral" not in service.tables:
-            raise RegTAPFeatureMissing("stc_spectral missing on current RegTAP service")
-
-        return super().get_search_condition(service)
+            raise RegTAPFeatureMissing(
+                "stc_spectral missing on current RegTAP service")
+        cond = super().get_search_condition(service)
+        if self.inclusive:
+            return (f"({cond}) OR NOT EXISTS("
+                "SELECT 1 FROM rr.stc_spectral AS inner_s WHERE"
+                    " inner_s.ivoid=rr.resource.ivoid)")
+        else:
+            return cond
 
 
 class Temporal(SubqueriedConstraint):
@@ -903,7 +942,7 @@ class Temporal(SubqueriedConstraint):
 
     takes_sequence = True
 
-    def __init__(self, times):
+    def __init__(self, times, *, inclusive=False):
         """
 
         Parameters
@@ -912,7 +951,14 @@ class Temporal(SubqueriedConstraint):
             A point in time or time interval to cover.  Plain numbers
             are interpreted as MJD.  All resources *overlapping* the
             interval are returned.
+
+        inclusive : bool, optional
+            Normally, this constraint will remove all resources that do
+            not declare their temproal coverage.  Pass inclusive=True to
+            retain these.
         """
+        self.inclusive = inclusive
+
         if isinstance(times, tuple):
             self._fillers = {
                 "time_lo": self._to_mjd(times[0]),
@@ -943,9 +989,16 @@ class Temporal(SubqueriedConstraint):
 
     def get_search_condition(self, service):
         if "rr.stc_temporal" not in service.tables:
-            raise RegTAPFeatureMissing("stc_temporal missing on current RegTAP service")
+            raise RegTAPFeatureMissing(
+                "stc_temporal missing on current RegTAP service")
 
-        return super().get_search_condition(service)
+        cond = super().get_search_condition(service)
+        if self.inclusive:
+            return (f"({cond}) OR NOT EXISTS("
+                "SELECT 1 FROM rr.stc_temporal AS inner_t WHERE"
+                    " inner_t.ivoid=rr.resource.ivoid)")
+        else:
+            return cond
 
 
 # NOTE: If you add new Contraint-s, don't forget to add them in
@@ -988,7 +1041,7 @@ def build_regtap_query(constraints, service):
         extra_tables |= set(constraint._extra_tables)
 
     joined_tables = ["rr.resource", "rr.capability", "rr.interface"
-                     ] + list(extra_tables)
+                     ] + list(sorted(extra_tables))
 
     # see comment in regtap.RegistryResource for the following
     # oddity